Replaced float by double in calculate barycentric coords method

+++ includes/Triangle.h
@@ -118,12 +118,12 @@ public:
 
   void calculateBarycentricCoords(double& u, double& v, double& w, const Point3D& point) const {
     const Vector3D v0 = this->v2.position - this->v1.position, v1 = this->v3.position - this->v1.position, v2 = point - this->v1.position;
-    float d00 = v0 * v0;
-    float d01 = v0 * v1;
-    float d11 = v1 * v1;
-    float d20 = v2 * v0;
-    float d21 = v2 * v1;
-    float denom = d00 * d11 - d01 * d01;
+    double d00 = v0 * v0;
+    double d01 = v0 * v1;
+    double d11 = v1 * v1;
+    double d20 = v2 * v0;
+    double d21 = v2 * v1;
+    double denom = d00 * d11 - d01 * d01;
     v = (d11 * d20 - d01 * d21) / denom;
     w = (d00 * d21 - d01 * d20) / denom;
     u = 1.0f - v - w;
